+++ src/index.ts
@@ -11,7 +11,14 @@ import "./styles/link.scss";
 // Hooks / Utils
 
 export type { PolymorphicProps, PolymorphicRef } from "./utils";
-export { cn, useFocusTrap, useReducedMotion, useSpringHover, useSpringPresence } from "./utils";
+export {
+	cn,
+	type FocusTrapOptions,
+	useFocusTrap,
+	useReducedMotion,
+	useSpringHover,
+	useSpringPresence,
+} from "./utils";
 export type {
 	ListboxItem,
 	UseListboxPopupArgs,